+++ spec-integration/integration.spec.js
@@ -1,6 +1,7 @@
-const { expect } = require('chai');
+require('dotenv').config();
+const chai = require('chai');
 const EventEmitter = require('events');
-const bunyan = require('bunyan');
+const logger = require('@elastic.io/component-commons-library/lib/logger/logger').getLogger();
 const sinon = require('sinon');
 const { AttachmentProcessor } = require('@elastic.io/component-commons-library');
 const Sftp = require('../lib/Sftp');
@@ -8,7 +9,10 @@ const deleteAction = require('../lib/actions/delete');
 const upload = require('../lib/actions/upload');
 const read = require('../lib/triggers/read');
 const lookupObject = require('../lib/actions/lookupObject');
-require('dotenv').config();
+const upsertFile = require('../lib/actions/upsertFile');
+
+const { expect } = chai;
+chai.use(require('chai-as-promised'));
 
 const PROCESSED_FOLDER_NAME = '.elasticio_processed';
 
@@ -18,7 +22,7 @@ class TestEmitter extends EventEmitter {
     this.data = [];
     this.end = 0;
     this.error = [];
-    this.logger = bunyan.createLogger({ name: 'dummy' });
+    this.logger = logger;
 
     this.on('data', (value) => this.data.push(value));
     this.on('error', (value) => this.error.push(value));
@@ -29,37 +33,55 @@ class TestEmitter extends EventEmitter {
 }
 
 // eslint-disable-next-line func-names
-describe('SFTP integration test - upload then download', function () {
-  this.timeout(20000000);
+describe('SFTP integration test - upload then download', () => {
   let sftp;
   let host;
   let username;
   let password;
   let port;
   let directory;
+  let cfg;
+  let sender;
+  let receiver;
   const testNumber = Math.floor(Math.random() * 10000);
 
-  before(() => {
-    if (!process.env.SFTP_HOSTNAME) { throw new Error('Please set SFTP_HOSTNAME env variable to proceed'); }
+  before(async () => {
+    if (!process.env.SFTP_HOSTNAME) {
+      throw new Error('Please set SFTP_HOSTNAME env variable to proceed');
+    }
     host = process.env.SFTP_HOSTNAME;
-    username = process.env.USERNAME;
-    password = process.env.PASSWORD;
+    username = process.env.SFTP_USER;
+    password = process.env.SFTP_PASSWORD;
     port = process.env.PORT;
-    directory = `/home/eiotesti/www/integration-test/test-${testNumber}/`;
-  });
-
-  it('Uploads attachment', async () => {
-    const cfg = {
+    directory = `/www/integration-test/test-${testNumber}/`;
+    cfg = {
       host,
       username,
       password,
       port,
       directory,
     };
-    sftp = new Sftp(bunyan.createLogger({ name: 'dummy' }), cfg);
+    sftp = new Sftp(logger, cfg);
     await sftp.connect();
+  });
 
-    const sender = new TestEmitter();
+  beforeEach(async () => {
+    cfg = {
+      host,
+      username,
+      password,
+      port,
+      directory,
+    };
+    sender = new TestEmitter();
+    receiver = new TestEmitter();
+  });
+
+  after(async () => {
+    await sftp.end();
+  });
+
+  it('Uploads attachment', async () => {
     const msg = {
       body: {},
       attachments: {
@@ -82,16 +104,6 @@ describe('SFTP integration test - upload then download', function () {
   });
 
   it('Uploads and reads attachments', async () => {
-    const cfg = {
-      host,
-      username,
-      password,
-      port,
-      directory,
-    };
-    sftp = new Sftp(bunyan.createLogger({ name: 'dummy' }), cfg);
-    await sftp.connect();
-
     await upload.process.call(new TestEmitter(), {
       body: {},
       attachments: {
@@ -104,13 +116,12 @@ describe('SFTP integration test - upload then download', function () {
       },
     }, cfg);
 
-    const receiver = new TestEmitter();
     const msg = {};
     await read.process.call(receiver, msg, cfg);
     expect(receiver.data.length).to.equal(2);
-    expect(receiver.data[0].body.filename).to.equal('logo.svg');
+    expect(receiver.data[0].body.filename).to.equal('logo2.svg');
     expect(receiver.data[0].body.size).to.equal(4379);
-    expect(receiver.data[1].body.filename).to.equal('logo2.svg');
+    expect(receiver.data[1].body.filename).to.equal('logo.svg');
     expect(receiver.data[1].body.size).to.equal(4379);
     const logoFilename = (await sftp.list(`${cfg.directory}${PROCESSED_FOLDER_NAME}`))[0].name;
     const logo2Filename = (await sftp.list(`${cfg.directory}${PROCESSED_FOLDER_NAME}`))[1].name;
@@ -121,16 +132,6 @@ describe('SFTP integration test - upload then download', function () {
   });
 
   it('Uploads and reads attachments with custom name', async () => {
-    const cfg = {
-      host,
-      username,
-      password,
-      port,
-      directory,
-    };
-    sftp = new Sftp(bunyan.createLogger({ name: 'dummy' }), cfg);
-    await sftp.connect();
-
     await upload.process.call(new TestEmitter(), {
       body: { filename: 'custom.svg' },
       attachments: {
@@ -143,7 +144,6 @@ describe('SFTP integration test - upload then download', function () {
       },
     }, cfg);
 
-    const receiver = new TestEmitter();
     const msg = {};
     await read.process.call(receiver, msg, cfg);
     expect(receiver.data.length).to.equal(2);
@@ -160,16 +160,6 @@ describe('SFTP integration test - upload then download', function () {
   });
 
   it('Uploads, read and deletes attachments with custom name', async () => {
-    const cfg = {
-      host,
-      username,
-      password,
-      port,
-      directory,
-    };
-    sftp = new Sftp(bunyan.createLogger({ name: 'dummy' }), cfg);
-    await sftp.connect();
-
     await upload.process.call(new TestEmitter(), {
       body: { filename: 'custom.svg' },
       attachments: {
@@ -182,7 +172,6 @@ describe('SFTP integration test - upload then download', function () {
       },
     }, cfg);
 
-    const receiver = new TestEmitter();
     const msg = {};
     await read.process.call(receiver, msg, cfg);
     expect(receiver.data.length).to.equal(2);
@@ -208,17 +197,6 @@ describe('SFTP integration test - upload then download', function () {
   });
 
   it('Uploads, reads, and filters files by pattern match', async () => {
-    const cfg = {
-      host,
-      username,
-      password,
-      port,
-      directory,
-      pattern: 'pattern',
-    };
-    sftp = new Sftp(bunyan.createLogger({ name: 'dummy' }), cfg);
-    await sftp.connect();
-
     await upload.process.call(new TestEmitter(), {
       body: {},
       attachments: {
@@ -233,15 +211,14 @@ describe('SFTP integration test - upload then download', function () {
 
     const list = await sftp.list(cfg.directory);
     expect(list.length).to.equal(2);
-    expect(list[0].name).to.equal('logo.svg');
+    expect(list[0].name).to.equal('pattern.svg');
     expect(list[0].size).to.equal(4379);
-    expect(list[1].name).to.equal('pattern.svg');
+    expect(list[1].name).to.equal('logo.svg');
     expect(list[1].size).to.equal(4379);
 
-    const receiver = new TestEmitter();
     const msg = {};
+    cfg.pattern = 'pattern*';
     await read.process.call(receiver, msg, cfg);
-
     expect(receiver.data.length).to.equal(1);
     expect(receiver.data[0].body.filename).to.equal('pattern.svg');
     expect(receiver.data[0].body.size).to.equal(4379);
@@ -255,15 +232,6 @@ describe('SFTP integration test - upload then download', function () {
   it('Uploads and lookup', async () => {
     const attachmentProcessorStub = sinon.stub(AttachmentProcessor.prototype, 'uploadAttachment');
     const callAttachmentProcessor = attachmentProcessorStub.returns({ config: { url: 'https://url' } });
-    const cfg = {
-      host,
-      username,
-      password,
-      port,
-      directory,
-    };
-    sftp = new Sftp(bunyan.createLogger({ name: 'dummy' }), cfg);
-    await sftp.connect();
 
     await upload.process.call(new TestEmitter(), {
       body: {
@@ -280,7 +248,6 @@ describe('SFTP integration test - upload then download', function () {
     expect(list.length).to.equal(1);
     expect(list[0].name).to.equal('logo.svg');
 
-    const receiver = new TestEmitter();
     const msg = {
       body: {
         path: `${directory}/logo.svg`,
@@ -294,7 +261,152 @@ describe('SFTP integration test - upload then download', function () {
     attachmentProcessorStub.restore();
   });
 
-  afterEach(async () => {
-    await sftp.end();
+  describe('Upsert File Tests', () => {
+    // eslint-disable-next-line max-len
+    const attachmentUrl1 = 'https://gist.githubusercontent.com/jhorbulyk/1bc92d62a7a530ce19c83a4f5b7a9f88/raw/62ab4e2e3028315c3472d622a279f13eb44c4f44/Hello%2520World%2520Gist';
+    const attachmentUrl1ContentSize = 11;
+    // eslint-disable-next-line max-len
+    const attachmentUrl2 = 'https://gist.githubusercontent.com/jhorbulyk/e6ff536dac6e0caa1b2d2d0177e72855/raw/77b39d3cd443128becae7f1c1ed88950cd29e576/Hello%2520World%2520Gist%25202';
+    const attachmentUrl2ContentSize = 15;
+    let filename;
+    beforeEach(() => {
+      filename = `${directory}test.file`;
+    });
+
+    it('Relative Path Test', async () => {
+      cfg = {
+        host,
+        username,
+        password,
+        port,
+        updateBehavior: 'error',
+      };
+
+      sender = new TestEmitter();
+      const msg = {
+        body: {
+          filename: filename.replace('/home/eiotesti/', './'),
+          attachmentUrl: attachmentUrl1,
+        },
+      };
+      const result = await upsertFile.process.call(sender, msg, cfg);
+
+      expect(result.body.size).to.equal(attachmentUrl1ContentSize);
+      const list = await sftp.list(directory);
+      expect(list.length).to.equal(1);
+      expect(list[0].name).to.equal('test.file');
+      expect(list[0].size).to.equal(attachmentUrl1ContentSize);
+    });
+
+
+    it('Error Mode', async () => {
+      cfg = {
+        host,
+        username,
+        password,
+        port,
+        updateBehavior: 'error',
+      };
+
+      sender = new TestEmitter();
+      const msg = {
+        body: {
+          filename,
+          attachmentUrl: attachmentUrl1,
+        },
+      };
+      const result = await upsertFile.process.call(sender, msg, cfg);
+
+      expect(result.body.size).to.equal(attachmentUrl1ContentSize);
+      const list = await sftp.list(directory);
+      expect(list.length).to.equal(1);
+      expect(list[0].name).to.equal('test.file');
+      expect(list[0].size).to.equal(attachmentUrl1ContentSize);
+
+      await expect(upsertFile.process.call(sender, msg, cfg)).to.be.rejectedWith(`File ${filename} exists. File updates are not permissible as per the current configuration.`);
+    });
+
+    it('Overwrite Mode', async () => {
+      cfg = {
+        host,
+        username,
+        password,
+        port,
+        updateBehavior: 'overwrite',
+      };
+
+      sender = new TestEmitter();
+      const msg1 = {
+        body: {
+          filename,
+          attachmentUrl: attachmentUrl1,
+        },
+      };
+      const result1 = await upsertFile.process.call(sender, msg1, cfg);
+
+      expect(result1.body.size).to.equal(attachmentUrl1ContentSize);
+      let list = await sftp.list(directory);
+      expect(list.length).to.equal(1);
+      expect(list[0].name).to.equal('test.file');
+      expect(list[0].size).to.equal(attachmentUrl1ContentSize);
+
+      const msg2 = {
+        body: {
+          filename,
+          attachmentUrl: attachmentUrl2,
+        },
+      };
+      const result2 = await upsertFile.process.call(sender, msg2, cfg);
+
+      expect(result2.body.size).to.equal(attachmentUrl2ContentSize);
+      list = await sftp.list(directory);
+      expect(list.length).to.equal(1);
+      expect(list[0].name).to.equal('test.file');
+      expect(list[0].size).to.equal(attachmentUrl2ContentSize);
+    });
+
+    it('Append Mode', async () => {
+      cfg = {
+        host,
+        username,
+        password,
+        port,
+        updateBehavior: 'append',
+      };
+
+      sender = new TestEmitter();
+      const msg1 = {
+        body: {
+          filename,
+          attachmentUrl: attachmentUrl1,
+        },
+      };
+      const result1 = await upsertFile.process.call(sender, msg1, cfg);
+
+      expect(result1.body.size).to.equal(attachmentUrl1ContentSize);
+      let list = await sftp.list(directory);
+      expect(list.length).to.equal(1);
+      expect(list[0].name).to.equal('test.file');
+      expect(list[0].size).to.equal(attachmentUrl1ContentSize);
+
+      const msg2 = {
+        body: {
+          filename,
+          attachmentUrl: attachmentUrl2,
+        },
+      };
+      const result2 = await upsertFile.process.call(sender, msg2, cfg);
+
+      expect(result2.body.size).to.equal(attachmentUrl1ContentSize + attachmentUrl2ContentSize);
+      list = await sftp.list(directory);
+      expect(list.length).to.equal(1);
+      expect(list[0].name).to.equal('test.file');
+      expect(list[0].size).to.equal(attachmentUrl1ContentSize + attachmentUrl2ContentSize);
+    });
+
+    afterEach(async () => {
+      await sftp.delete(filename);
+      await sftp.rmdir(directory, false);
+    });
   });
 });