+++ spec-integration/pollingIntegration.spec.js
@@ -1,11 +1,12 @@
+require('dotenv').config();
+
 const { expect } = require('chai');
 const EventEmitter = require('events');
-const bunyan = require('bunyan');
+const logger = require('@elastic.io/component-commons-library/lib/logger/logger').getLogger();
 const nock = require('nock');
 const Sftp = require('../lib/Sftp');
 const upload = require('../lib/actions/upload');
 const poll = require('../lib/triggers/polling');
-require('dotenv').config();
 
 class TestEmitter extends EventEmitter {
   constructor() {
@@ -13,7 +14,7 @@ class TestEmitter extends EventEmitter {
     this.data = [];
     this.end = 0;
     this.error = [];
-    this.logger = bunyan.createLogger({ name: 'dummy' });
+    this.logger = logger;
 
     this.on('data', (value) => this.data.push(value));
     this.on('error', (value) => this.error.push(value));
@@ -24,23 +25,36 @@ class TestEmitter extends EventEmitter {
 }
 
 // eslint-disable-next-line func-names
-describe('SFTP integration test - polling', function () {
-  this.timeout(2000000);
+describe('SFTP integration test - polling', () => {
   let sftp;
+  let cfg;
   let host;
   let username;
   let password;
   let port;
   let directory;
+  let sender;
   const testNumber = Math.floor(Math.random() * 10000);
 
-  before(() => {
-    if (!process.env.SFTP_HOSTNAME) { throw new Error('Please set SFTP_HOSTNAME env variable to proceed'); }
+  before(async () => {
+    if (!process.env.SFTP_HOSTNAME) {
+      throw new Error('Please set SFTP_HOSTNAME env variable to proceed');
+    }
     host = process.env.SFTP_HOSTNAME;
-    username = process.env.USERNAME;
-    password = process.env.PASSWORD;
+    username = process.env.SFTP_USER;
+    password = process.env.SFTP_PASSWORD;
     port = process.env.PORT;
-    directory = `/home/eiotesti/www/integration-test/test-${testNumber}/`;
+    directory = `/www/integration-test/test-${testNumber}/`;
+    cfg = {
+      host,
+      username,
+      password,
+      port,
+      directory,
+    };
+    sftp = new Sftp(logger, cfg);
+    await sftp.connect();
+    sender = new TestEmitter();
   });
 
   it('Uploads and poll attachment', async () => {
@@ -50,26 +64,16 @@ describe('SFTP integration test - polling', function () {
     nock('http://api.io/', { encodedQueryParams: true })
       .put('/some').reply(200, { signedUrl: { put_url: 'http://api.io/some' } });
 
-    const cfg = {
-      host,
-      username,
-      password,
-      port,
-      directory,
-    };
-    sftp = new Sftp(bunyan.createLogger({ name: 'dummy' }), cfg);
-    await sftp.connect();
 
-    const sender = new TestEmitter();
     const msg = {
-      body: {},
+      body: { filename: 'logo.svg' },
       attachments: {
         'logo.svg': {
           url: 'https://app.elastic.io/img/logo.svg',
         },
       },
     };
-    const result = await upload.process.call(sender, msg, cfg);
+    const result = await upload.process.call(new TestEmitter(), msg, cfg);
 
     expect(result.body.results).to.be.an('array');
     expect(result.body.results.length).to.equal(1);
@@ -78,7 +82,6 @@ describe('SFTP integration test - polling', function () {
     expect(list.length).to.equal(1);
     expect(list[0].name).to.equal('logo.svg');
     expect(list[0].size).to.equal(4379);
-
     await poll.process.call(sender, {}, cfg);
 
     expect(sender.data[0].body.path).to.equal(`${cfg.directory}logo.svg`);
@@ -88,7 +91,7 @@ describe('SFTP integration test - polling', function () {
     await sftp.rmdir(cfg.directory, false);
   });
 
-  afterEach(async () => {
+  after(async () => {
     await sftp.end();
   });
 });