+++ spec/attachments.spec.js
@@ -15,20 +15,38 @@ const stream = new Stream();
 const contentLength = 10;
 
 describe('Attachment tests', () => {
+  let uploadAttachment;
+  before(() => {
+  });
   afterEach(() => {
     self.emit.resetHistory();
   });
+  after(() => {
+    uploadAttachment.restore();
+  });
 
   it('Adds an attachment correctly and returns the correct message', async () => {
-    const uploadAttachment = sinon.stub(AttachmentProcessor.prototype, 'uploadAttachment').resolves(result);
+    uploadAttachment = sinon.stub(AttachmentProcessor.prototype, 'uploadAttachment').resolves(result);
     await attachments.addAttachment.call(self, msg, name, stream, contentLength);
     expect(uploadAttachment.calledOnceWithExactly(stream, 'stream')).to.be.equal(true);
     expect(msg).to.be.deep.equal({ attachments: { file: { url: '/hello/world', size: 10 } } });
     uploadAttachment.restore();
   });
 
+  it('Throws an error if attachment file is too large', async () => {
+    const file = {
+      type: '-',
+      name: '1.txt',
+      size: 70000000000,
+      accessTime: '1575379317000',
+      modifyTime: '1575291942000',
+    };
+    await attachments.addAttachment.call(self, msg, file.name, stream, file.size);
+    expect(self.emit.getCall(0).args[1].message).to.be.equal('File is 70000000000 bytes, and is too large to upload as an attachment. Max attachment size is 10485760 bytes');
+  });
+
   it('Emits an error upon failure', async () => {
-    const uploadAttachment = sinon.stub(AttachmentProcessor.prototype, 'uploadAttachment').throws(new Error('This input should be rejected'));
+    uploadAttachment = sinon.stub(AttachmentProcessor.prototype, 'uploadAttachment').throws(new Error('This input should be rejected'));
 
     await attachments.addAttachment.call(self, msg, name, 'not a stream', contentLength)
       .catch((e) => {